Renaming screen to driver
