+++ vaitk/core/drivers/abc/abc_driver.py
@@ -1,21 +1,40 @@
-from abc import abstractmethod
-from vaitk.gui.AbstractRectangularArea import AbstractRectangularArea
+from abc import abstractmethod, ABCMeta
 
 
-class AbstractScreen(AbstractRectangularArea):
+class ABCDriver(metaclass=ABCMeta):
+    """
+    ABC for the interface of a generic driver to a input/output device.
+    """
+
+    @abstractmethod
+    def init(self):
+        """Initialises the device"""
+
+    @abstractmethod
+    def deinit(self):
+        """Deinitialises the device"""
+
+    @property
     @abstractmethod
     def num_colors(self):
         """
         Returns the supported number of colors
         """
 
+    @property
     @abstractmethod
     def cursor_pos(self):
         """Gets the current position of the cursor"""
 
+    @cursor_pos.setter
     @abstractmethod
-    def set_cursor_pos(self, pos):
-        """Sets the position of the cursor"""
+    def cursor_pos(self, pos):
+        """Set the current position of the cursor"""
+
+    @property
+    @abstractmethod
+    def size(self):
+        """Returns the size of the accessible area."""
 
     @abstractmethod
     def reset(self):
@@ -27,3 +46,8 @@ class AbstractScreen(AbstractRectangularArea):
         Writes a string at a given position and with given foreground
         and background colors
         """
+
+    @abstractmethod
+    def get_key_code(self):
+        """Gets the keycode from a device. This method must be blocking."""
+
