+++ vaitk/core/drivers/abc/abc_native_color.py
@@ -1,0 +1,4 @@
+import abc
+
+class ABCNativeColor(metaclass=abc.ABCMeta):
+    pass
