+++ vaitk/test/__init__.py
@@ -1,2 +1,1 @@
 from .SignalSpy import BaseSignalSpy  # noqa
-from .TextScreen import TextScreen  # noqa
